Fix bitwise operation between different enumeration types warning by casting to unsigned int

+++ llvm/lib/CodeGen/GlobalISel/InlineAsmLowering.cpp
@@ -52,7 +52,7 @@ public:
       Flags |= InlineAsm::Extra_IsAlignStack;
     if (CB.isConvergent())
       Flags |= InlineAsm::Extra_IsConvergent;
-    Flags |= IA->getDialect() * InlineAsm::Extra_AsmDialect;
+    Flags |= static_cast<unsigned int>(IA->getDialect()) * static_cast<unsigned int>(InlineAsm::Extra_AsmDialect);
   }
 
   void update(const TargetLowering::AsmOperandInfo &OpInfo) {
